+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2019-11
 
-- reflect feature update when edited fron coords-table
+- edit radius in coords-table
+- reflect feature update when edited from coords-table
 - include coords-table - preliminary
 - initial commit